Implement wrapper for phylophlan_write_config_file
Support the tools PhyloPhlAn can use out-of-the-box.
Encode the conditions for when parameters are required.

The generated configuration file contains a custom wrapping format for the tools
PhyloPhlAn needs and is intended to be edited by the user to extend support.
Direct access to this file would require deep knowledge of Galaxy internals and
allow the user to run arbitrary commands on the server. Alternatively, the
configuration step could be combined with the wrapper for the main phylophlan
script. In that case the wrapper would need to modify the scipt to add support
for further tools.

Todo:
Add the tools Phylophlan uses to the requirements section.
Investigate the threat model of galaxy.
- Integrate the configuration with the main scipt that uses it
  OR write tests if they can stay separate.