Initial wrapper file for phylophlan_assign_sgbs.py script
Implement most of the command and input selection.

This tool heavily relies on the included SGB databases, which have file sizes between 20-40GB.
It seems unrealistic than a user would ever create or upload their own database,
as the archive needs to adhere to a strict internal format.
These files definitely need to be cached using a data manager.
Additionally the --only_input option seems to be broken at the moment.