Add test to phylophlan_assign_sgbs
Fix loading the database from history.

It turns out that it's possible to delete a lot of the '.msh' mash files or replace them with smaller ones.
The tool will just report, that it could not find any close SGBs or that they are not in the database.
Most of the lines in the mapping file could also be deleted (but it will crash if a close SGB is missing in this file).